feat(api): remove Flask app from global scope
