+++ api/.gitignore
@@ -1,6 +1,7 @@
 .coverage
 coverage.xml
 
+*.log
 *.swp
 *.pyc
