+++ api/onnx_web/chain/__init__.py
@@ -13,3 +13,20 @@ from .source_txt2img import source_txt2img
 from .upscale_outpaint import upscale_outpaint
 from .upscale_resrgan import upscale_resrgan
 from .upscale_stable_diffusion import upscale_stable_diffusion
+
+CHAIN_STAGES = {
+    "blend-img2img": blend_img2img,
+    "blend-inpaint": blend_inpaint,
+    "blend-mask": blend_mask,
+    "correct-codeformer": correct_codeformer,
+    "correct-gfpgan": correct_gfpgan,
+    "persist-disk": persist_disk,
+    "persist-s3": persist_s3,
+    "reduce-crop": reduce_crop,
+    "reduce-thumbnail": reduce_thumbnail,
+    "source-noise": source_noise,
+    "source-txt2img": source_txt2img,
+    "upscale-outpaint": upscale_outpaint,
+    "upscale-resrgan": upscale_resrgan,
+    "upscale-stable-diffusion": upscale_stable_diffusion,
+}
